types arrangement and style, extract chatc onversation
