+++ src/types/Conversation.ts
@@ -1,0 +1,4 @@
+export type Conversation = {
+    message: string;
+    timestamp: number;
+  };
