+++ src/types/ConversationMap.ts
@@ -1,0 +1,6 @@
+import { Conversation } from ".";
+
+export type ConversationMap = {
+  email: string;
+  conversation: Array<Conversation>;
+};
