+++ src/types/Friend.ts
@@ -1,0 +1,4 @@
+export type Friend = {
+    email: string;
+    name: string;
+  };
